+++ Src/Matrix.fs
@@ -5,6 +5,9 @@ open System.Runtime.CompilerServices // for [<IsByRefLike; IsReadOnly>] see http
 open Euclid.UtilEuclid
 open System.Runtime.Serialization // for serialization of struct fields only but not properties via  [<DataMember>] attribute. with Newtonsoft.Json or similar
 open EuclidErrors
+#if !FABLE_COMPILER
+open System.Text.Json.Serialization
+#endif
 
 
 
@@ -28,9 +31,12 @@ open EuclidErrors
 [<Struct; NoEquality; NoComparison>] // because its made up from floats
 [<IsReadOnly>]
 //[<IsByRefLike>]
-[<DataContract>] // for using DataMember on fields
+[<DataContract>] // for using DataMember on fields, for Newtonsoft.Json
+#if !FABLE_COMPILER
+[<JsonConverter(typeof<MatrixJsonConverter>)>]
+#endif
 type Matrix =
-    //[<DataMember>] //to serialize this struct field (but not properties) with Newtonsoft.Json and similar
+    
     /// The element in row 1, column 1 of the matrix.
     [<DataMember>] val M11 : float
     /// The element in row 1, column 2 of the matrix.
@@ -1251,3 +1257,25 @@ type Matrix =
                 )
     *)
 
+#if !FABLE_COMPILER
+/// Serializes a Matrix as its 16 elements with System.Text.Json.
+and MatrixJsonConverter() =
+    inherit JsonConverter<Matrix>()
+
+    let names =
+        [| "M11"; "M21"; "M31"; "X41"
+           "M12"; "M22"; "M32"; "Y42"
+           "M13"; "M23"; "M33"; "Z43"
+           "M14"; "M24"; "M34"; "M44" |]
+
+    override _.Write(writer, matrix, options) =
+        JsonConvert.writeFloatProperties writer options "Euclid.Matrix" names
+            [| matrix.M11; matrix.M21; matrix.M31; matrix.X41
+               matrix.M12; matrix.M22; matrix.M32; matrix.Y42
+               matrix.M13; matrix.M23; matrix.M33; matrix.Z43
+               matrix.M14; matrix.M24; matrix.M34; matrix.M44 |]
+
+    override _.Read(reader, _, options) =
+        let v = JsonConvert.readFloatProperties &reader options "Euclid.Matrix" names
+        Matrix.createFromRowMajorArray v
+#endif